Add conditional check for opening cash drawer
Added a comment linking the cash drawer feature to Feature #3697. Implemented a conditional check to determine if the cash drawer should be opened based on a system configuration value. If the configuration indicates to open the drawer, a log entry is created and the `bt_OpenCashDrawer` button's click event is triggered.